+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+# 1.3.1
+- little improvement in `drawDefaultLifeBar`.
+- create mixin `AutomaticRandomMovement`
+- add `onReady` in `BonfireTiledWidget` and `BonfireWidget`
+- add `getComponentDirectionFromMe` in `GameComponentExtensions`
+- add `checkInterval` in `GameComponent`
+
 # 1.3.0
 - new extensions to `GameComponent`.
 - new extensions to `Movement`.